requirements.txt
torch>=2.2
PyYAML>=6.0
numpy>=1.24
tokenizers>=0.15
sentencepiece>=0.1.99
